+++ src/pages/Metadata/MetadataExamples.vue
@@ -2,38 +2,24 @@
   <div class="api-producer-metadata-examples">
     <ContentHeader
         title="Metadata Examples"
-        introduction="Your API has some technical preferences and architectural decisions that needs to be convaied in order to enable other developer to reuse the functionality." />
+        introduction="Your API has some technical preferences and architectural decisions that needs to be communicated in order to enable other developer to reuse the functionality." />
     <Content>
       <CommandLine
-          title="“Design and architecuture” phase"
+          title="“Design and architecture” phase"
           :command="exampleOne()" />
       <p>There is a need for new functionality and the team has just started developing the solution architecture. With external customers being the primary users there is a need to customize the response body.</p>
-      <h5>Technology</h5>
-      <p>This is a JSON Rest API</p>
-      <h5>API-First strategy</h5>
-      <p>Design First</p>
     </Content>
     <Content>
       <CommandLine
-          title="Off-the-shelf solution has been extended with new functionallity"
+          title="Off-the-shelf solution has been extended with new functionality"
           :command="exampleTwo()" />
-      <p>
-        A previously purchased software solution has been updated and got some new functionality that could be useful to the team and should not be used utsiden of the organization.
-      </p>
-      <h5>Technology</h5>
-      <p>This is a SOAP API</p>
-      <h5>API-First strategy</h5>
-      <p>None, off-the-shelf solution purchased</p>
+      <p>A previously purchased software solution has been updated and got some new functionality that could be useful to the team and should not be used outside of the organization.</p>
     </Content>
     <Content>
       <CommandLine
           title="Deprecated API"
           :command="exampleFour()" />
-      <p>Our old API has been deprecated and should no longer be used by new clients. Since other services allready uses it, the documentation data must be available.</p>
-      <h5>Technology</h5>
-      <p>This is a JSON Rest API</p>
-      <h5>API-First strategy</h5>
-      <p>None</p>
+      <p>Our old API has been deprecated and should no longer be used by new clients. Since other services already uses it, the documentation data must be available.</p>
     </Content>
     <Footer />
   </div>
@@ -43,9 +29,8 @@
 
 import Footer from "@/components/Elements/Footer";
 import CommandLine from "@apimap/command-line";
-import {  Content, ContentHeader, CenterLayout } from "@apimap/layout-core"
+import { Content, ContentHeader, CenterLayout } from "@apimap/layout-core"
 import { MetadataExamples } from "@/data/MetadataExamples";
-import {TaxonomyExamples} from "@/data/TaxonomyExamples";
 
 export default {
   name: "MetadataUseCases",